deployment/requirements.txt
pandas==2.2.2
streamlit==1.44.0
numpy
seaborn
plotly
matplotlib
pillow
scikit-learn==1.5.1
joblib==1.5.1